requirements.txt
pycaw
pystray
Pillow
psutil
pywin32
comtypes
